Re-export hash_map + hash_set modules rather than renaming
