+++ crates/bevy_ecs/src/relationship/relationship_source_collection.rs
@@ -1,4 +1,4 @@
-use crate::entity::{Entity, EntityHashSet};
+use crate::entity::{hash_set::EntityHashSet, Entity};
 use alloc::vec::Vec;
 use smallvec::SmallVec;
 
@@ -65,7 +65,7 @@ impl RelationshipSourceCollection for Vec<Entity> {
 }
 
 impl RelationshipSourceCollection for EntityHashSet {
-    type SourceIter<'a> = core::iter::Copied<crate::entity::EntityHashSetIter<'a>>;
+    type SourceIter<'a> = core::iter::Copied<crate::entity::hash_set::EntityHashSetIter<'a>>;
 
     fn with_capacity(capacity: usize) -> Self {
         EntityHashSet::with_capacity(capacity)
