+++ crates/bevy_pbr/src/cluster/mod.rs
@@ -5,7 +5,7 @@ use core::num::NonZero;
 use bevy_core_pipeline::core_3d::Camera3d;
 use bevy_ecs::{
     component::Component,
-    entity::{Entity, EntityHashMap},
+    entity::{Entity, hash_map::EntityHashMap},
     query::{With, Without},
     reflect::ReflectComponent,
     system::{Commands, Query, Res, Resource},
