+++ crates/bevy_pbr/src/meshlet/resource_manager.rs
@@ -7,7 +7,7 @@ use bevy_core_pipeline::{
 };
 use bevy_ecs::{
     component::Component,
-    entity::{Entity, EntityHashMap},
+    entity::{Entity, hash_map::EntityHashMap},
     query::AnyOf,
     system::{Commands, Query, Res, ResMut, Resource},
 };
