+++ examples/tools/scene_viewer/animation_plugin.rs
@@ -1,7 +1,9 @@
 //! Control animations of entities in the loaded scene.
 use std::collections::HashMap;
 
-use bevy::{animation::AnimationTarget, ecs::entity::EntityHashMap, gltf::Gltf, prelude::*};
+use bevy::{
+    animation::AnimationTarget, ecs::entity::hash_map::EntityHashMap, gltf::Gltf, prelude::*,
+};
 
 use crate::scene_viewer_plugin::SceneHandle;
